### 2024-04-02 #6572 @editorialbot
[PRE REVIEW]: SCIMAP: A Python Toolkit for Integrated Spatial Analysis of Multiplexed Imaging Data
**Submitting author:** @ajitjohnson (<a href="http://orcid.org/0000-0003-4139-0618">Ajit Johnson Nirmal</a>)
**Repository:** https://github.com/labsyspharm/scimap
**Branch with paper.md** (empty if default branch):
**Version:** v2.0.0
