+++ paper/paper.md
@@ -25,7 +25,7 @@ bibliography: paper.bib
 
 # Summary
 
-Multiplexed imaging data are revolutionizing our understanding of the composition and organization of tissues and tumors [@noauthor_catching_2022]. A critical aspect of such “tissue profiling” is quantifying the spatial relationship relationships among cells at different scales from the interaction of neighboring cells to recurrent communities of cells of multiple types. This often involves statistical analysis of 10<sup>7</sup> or more cells in which up to 100 biomolecules (commonly proteins) have been measured. While software tools currently cater to the analysis of spatial transcriptomics data [@liu_analysis_2022], there remains a need for toolkits explicitly tailored to the complexities of multiplexed imaging data including the need to seamlessly integrate image visualization with data analysis and exploration. We introduce `SCIMAP`, a Python package specifically crafted to address these challenges. With `SCIMAP`, users can efficiently preprocess, analyze, and visualize large datasets, facilitating the exploration of spatial relationships and their statistical significance. `SCIMAP`'s modular design enables the integration of new algorithms, enhancing its capabilities for spatial analysis.
+Multiplexed imaging data are revolutionizing our understanding of the composition and organization of tissues and tumors [@noauthor_catching_2022]. A critical aspect of such “tissue profiling” is quantifying the spatial relationships among cells at different scales from the interaction of neighboring cells to recurrent communities of cells of multiple types. This often involves statistical analysis of 10$^7$ or more cells in which up to 100 biomolecules (commonly proteins) have been measured. While software tools currently cater to the analysis of spatial transcriptomics data [@liu_analysis_2022], there remains a need for toolkits explicitly tailored to the complexities of multiplexed imaging data including the need to seamlessly integrate image visualization with data analysis and exploration. We introduce `SCIMAP`, a Python package specifically crafted to address these challenges. With `SCIMAP`, users can efficiently preprocess, analyze, and visualize large datasets, facilitating the exploration of spatial relationships and their statistical significance. `SCIMAP`'s modular design enables the integration of new algorithms, enhancing its capabilities for spatial analysis.
 
 # Statement of Need
 
@@ -33,7 +33,7 @@ A variety of methods have been introduced for high multiplexed imaging of tissue
 
[…]
