### 2024-04-02 #6572 @ajitjohnson
@csoneson It was very nice to meet you. Thank you for pointing out those errors. I have corrected them and also replaced `<sup>` with the `10$^7$` notation.
