### 2024-04-02 81bab14
removed an extra } in figure legends
