+++ paper/paper.md
@@ -39,7 +39,7 @@ Lastly, a single high-plex whole slide image can exceed 100GB per image and 10$^
 
 `SCIMAP` is available as a standalone Python package for interactive use, in Jupyter Notebook for example, or can be accessed via a command-line interface (CLI; only a subset of functions that do not require visualization) for cloud-based processing. The package can be accessed at [github](https://github.com/labsyspharm/scimap) and installed locally through pip. Installation, usage instructions, general documentation, and tutorials, are available at [https://scimap.xyz/](https://scimap.xyz/). See See \autoref{fig:workflow} for a schematic of the workflow and system components.
 
-![SCIMAP Workflow Overview. The schematic highlights data import, cell classification, spatial analysis, and visualization techniques within the SCIMAP tool box.\label{fig:workflow}}](figure-workflow.png)
+![SCIMAP Workflow Overview. The schematic highlights data import, cell classification, spatial analysis, and visualization techniques within the SCIMAP tool box.\label{fig:workflow}](figure-workflow.png)
 
 `SCIMAP` comprises of four main modules: preprocessing, analysis tools, visualization, and external methods. The preprocessing tools include functions for normalization, batch correction, and streamlined import from cloud processing pipelines such as MCMICRO [@schapiro_mcmicro_2022]. The analysis tools offer standard single-cell analysis techniques such as dimensionality reduction, clustering, prior knowledge-based cell phenotyping (a method through which cells are classified into specific cell types based on patterns of marker expression defined by the user), and various spatial analysis tools for measuring cellular distances, identifying regions of specific cell type aggregation, and assessing statistical differences in proximity scores or interaction frequencies. `SCIMAP` also includes neighborhood detection algorithms that utilize spatial-LDA [@wang_spatial_2007] for categorical data (cell types or clusters) and spatial lag for continuous data (marker expression values). Most analysis tools come with corresponding visualization functions to plot the results effectively. Additionally, the external methods module facilitates the integration of new tools developed by the community into `SCIMAP`, further extending its utility and applicability to both 2D and 3D data. 
 
