+++ pyproject.toml
@@ -1,7 +1,7 @@
 [tool.poetry]
 
 name = "SCIMAP"
-version = "2.0.0"
+version = "2.0.1"
 description = "Spatial Single-Cell Analysis Toolkit"
 
 license = "MIT"
@@ -44,17 +44,19 @@ plotly = "^5.18.0"
 TiffFile = "^2023.9.26"
 dask = {extras = ["array"], version = "^2023.11.0"}
 zarr = "2.10.3"
-napari = "^0.4.18"
+napari = {version = "^0.4.18", optional = true}
 numba = ">=0.58.1"
 shapely = "^2.0.2"
 gensim = "^4.3.2"
 mkdocs-material = "^9.4.8"
-napari-ome-zarr = "^0.4.0"
+napari-ome-zarr = {version = "^0.4.0", optional = true}
 llvmlite = "^0.41.1"
 combat = "^0.3.3"
 mpl-scatter-density = "^0.7"
-pyqt6 = "^6.6.1"
+pyqt6 = {version = "^6.6.1", optional = true}
 
+[tool.poetry.extras]
+napari = ["pyqt6", "napari-ome-zarr", "napari"]
 
 [tool.poetry.dev-dependencies]
 
