+++ scimap/plotting/gate_finder.py
@@ -27,8 +27,8 @@ from dask.cache import Cache
 import zarr
 import os
 
-cache = Cache(2e9)  # Leverage two gigabytes of memory
-cache.register()
+# cache = Cache(2e9)  # Leverage two gigabytes of memory
+# cache.register()
 
 
 def gate_finder(
@@ -52,84 +52,84 @@ def gate_finder(
     **kwargs,
 ):
     """
-Parameters:
-        image_path (str):
-            Path to the high-resolution image file (supports formats like TIFF, OME.TIFF).
+    Parameters:
+            image_path (str):
+                Path to the high-resolution image file (supports formats like TIFF, OME.TIFF).
 
-        adata (anndata.AnnData):
-            The annotated data matrix.
+            adata (anndata.AnnData):
+                The annotated data matrix.
 
-        marker_of_interest (str):
-            The target marker for which the gating threshold is to be determined.
+            marker_of_interest (str):
+                The target marker for which the gating threshold is to be determined.
 
-        layer (str, optional):
-            Specifies the layer in `adata` containing expression data. Defaults to 'raw' for `adata.raw.X`.
+            layer (str, optional):
+                Specifies the layer in `adata` containing expression data. Defaults to 'raw' for `adata.raw.X`.
 
-        log (bool, optional):
-            Applies log transformation to expression data if set to True. 
+            log (bool, optional):
+                Applies log transformation to expression data if set to True.
 
-        from_gate (int, optional):
-            Starting gate threshold value for the marker of interest. 
+            from_gate (int, optional):
+                Starting gate threshold value for the marker of interest.
 
-        to_gate (int, optional):
-            Ending gate threshold value for the marker of interest. 
+            to_gate (int, optional):
+                Ending gate threshold value for the marker of interest.
 
-        increment (float, optional):
-            Incremental step size between `from_gate` and `to_gate`. 
+            increment (float, optional):
+                Incremental step size between `from_gate` and `to_gate`.
 
-        markers (list, optional):
-            A list of additional markers to include in visualization for context.
+            markers (list, optional):
+                A list of additional markers to include in visualization for context.
 
-        channel_names (list or str, optional):
-            Names of the channels in the image, in order. Defaults to 'default', using `adata.uns['all_markers']`.
+            channel_names (list or str, optional):
+                Names of the channels in the image, in order. Defaults to 'default', using `adata.uns['all_markers']`.
 
-        flip_y (bool, optional):
-            Inverts the Y-axis to match image coordinates if set to True. Defaults to True.
+            flip_y (bool, optional):
+                Inverts the Y-axis to match image coordinates if set to True. Defaults to True.
 
-        x_coordinate, y_coordinate (str, optional):
-            Columns in `adata.obs` specifying cell coordinates. Defaults are 'X_centroid' and 'Y_centroid'.
+            x_coordinate, y_coordinate (str, optional):
+                Columns in `adata.obs` specifying cell coordinates. Defaults are 'X_centroid' and 'Y_centroid'.
 
-        point_size (int, optional):
-            Size of points in the visualization. 
+            point_size (int, optional):
+                Size of points in the visualization.
 
-        imageid (str, optional):
-            Column in `adata.obs` identifying images for datasets with multiple images. 
+            imageid (str, optional):
+                Column in `adata.obs` identifying images for datasets with multiple images.
 
-        subset (str, optional):
[…]
