+++ scimap/plotting/addROI_image.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
-#Created on Mon Apr  4 11:12:59 2022
-#@author: Ajit Johnson Nirmal
-#Adding ROI to data via Napari.
+# Created on Mon Apr  4 11:12:59 2022
+# @author: Ajit Johnson Nirmal
+# Adding ROI to data via Napari.
 
 """
 !!! abstract "Short Description"
@@ -26,7 +26,7 @@
 
 
 # Library
-#%gui qt
+# %gui qt
 try:
     import napari
     from napari.layers import Shapes
@@ -47,139 +47,140 @@ import scipy.spatial.distance as sdistance
 from joblib import Parallel, delayed
 
 
-
-cache = Cache(2e9)  # Leverage two gigabytes of memory
-cache.register()
+# cache = Cache(2e9)  # Leverage two gigabytes of memory
+# cache.register()
 
 
 # Function
-def addROI_image (image_path, 
-                  adata, 
-                  subset=None,
-                  imageid='imageid', 
-                  overlay=None, 
-                  flip_y=True,
-                  overlay_category=None,
-                  markers=None,
-                  channel_names='default',
-                  x_coordinate='X_centroid',
-                  y_coordinate='Y_centroid',
-                  point_size=10,
-                  point_color=None,
-                  seg_mask=None,
-                  n_jobs=-1, 
-                  verbose=False, 
-                  overwrite=True, 
-                  label='ROI', **kwargs):
+def addROI_image(
+    image_path,
+    adata,
+    subset=None,
+    imageid='imageid',
+    overlay=None,
+    flip_y=True,
+    overlay_category=None,
+    markers=None,
+    channel_names='default',
+    x_coordinate='X_centroid',
+    y_coordinate='Y_centroid',
+    point_size=10,
+    point_color=None,
+    seg_mask=None,
+    n_jobs=-1,
+    verbose=False,
+    overwrite=True,
+    label='ROI',
+    **kwargs,
+):
     """
-Parameters:
-        image_path (str):  
-            Path to the image file. Supports TIFF, OME.TIFF, and ZARR formats.
-            
-        adata (anndata.AnnData):  
-            The annotated data matrix to which ROIs will be added.
-            
-        subset (list, optional):  
-            Specifies the image(s) within `adata` to annotate. Necessary if `adata` contains multiple images.
-            
-        imageid (str, optional):  
-            Column in `adata.obs` identifying images, for datasets with multiple images. 
-            
-        overlay (str, optional):  
-            Column in `adata.obs` with categorical data to overlay on the image, such as cell phenotypes.
-            
-        flip_y (bool, optional):  
-            If True, inverts the Y-axis to match image coordinates. Default is True.
-            
-        overlay_category (list, optional):  
-            Specific categories within `overlay` to visualize. If None, all categories are shown.
-            
-        markers (list, optional):  
-            Specific markers to display. If None, all available markers are included.
-            
-        channel_names (list or str, optional):  
-            Names of image channels, in order as they appear in the image. Defaults to markers in `adata.uns['all_markers']` if 'default'.
-            
-        x_coordinate, y_coordinate (str, optional):  
-            Column names in `adata.obs` for cell coordinates. Defaults are 'X_centroid' and 'Y_centroid'.
-            
-        point_size (int, optional):  
-            Size of points in the visualization. 
-            
-        point_color (str, optional):  
-            Color of points in the visualization. If None, colors are automatically assigned.
-            
-        seg_mask (str, optional):  
-            Path to a segmentation mask file to be overlaid. 
-            
-        n_jobs (int, optional):  
-            Number of parallel jobs to use. Defaults to -1, using all available cores.
-            
-        verbose (bool, optional):  
[…]
