+++ pyproject.toml
@@ -1,7 +1,7 @@
 [tool.poetry]
 
 name = "SCIMAP"
-version = "2.0.2"
+version = "2.0.3"
 description = "Spatial Single-Cell Analysis Toolkit"
 
 license = "MIT"
