### 2024-04-09 #6572 @csoneson
👋🏻 @michaelberks, @NMontanaBrown - would one of you be interested in reviewing this submission for the Journal of Open Source Software (JOSS)?

> SCIMAP: A Python Toolkit for Integrated Spatial Analysis of Multiplexed Imaging Data
https://github.com/openjournals/joss-reviews/issues/6572

The review process happens on GitHub and is checklist-driven, see more [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html)

Thanks in advance!
Charlotte
