### 2024-04-10 f7bfe2d
scipy version update as the latest version does not work on windows
