+++ pyproject.toml
@@ -31,7 +31,7 @@ python = ">=3.9,<3.11"
 pytest = "^7.4.3"
 anndata = "^0.10.5.post1"
 pandas = "^2.1.3"
-scipy = "^1.11.3"
+scipy = ">1.11.3,<=1.12.0"
 seaborn = "^0.13.2"
 tifffile = "^2023.9.26"
 numpy = "^1.26.2"
