+++ pyproject.toml
@@ -1,7 +1,7 @@
 [tool.poetry]
 
 name = "SCIMAP"
-version = "2.0.4"
+version = "2.0.5"
 description = "Spatial Single-Cell Analysis Toolkit"
 
 license = "MIT"
