### 2024-04-11 #6604 @editorialbot
[REVIEW]: SCIMAP: A Python Toolkit for Integrated Spatial Analysis of Multiplexed Imaging Data
**Submitting author:** @ajitjohnson (<a href="http://orcid.org/0000-0003-4139-0618">Ajit Johnson Nirmal</a>)
**Repository:** https://github.com/labsyspharm/scimap
**Branch with paper.md** (empty if default branch):
**Version:** 2.1.0
