### 2024-04-11 #6604 @csoneson
👋🏼 @ajitjohnson, @jluethi, @kevinyamauchi, @NMontanaBrown - this is the review thread for the submission. All of our communications will happen here from now on.

```@editorialbot generate my checklist```

as the top of a new comment in this thread. These checklists contain the JOSS requirements. As you go over the submission, please check any items that you feel have been satisfied. The first comment in this thread also contains links to the JOSS reviewer guidelines.
