### 2024-04-15 labsyspharm/scimap#98 @kevinyamauchi
.vscode in the repo
Hello! Super minor point, but I noticed that a .vscode directory is included in the repo. It would be nice to remove as these  VSCode settings might collide with other people's settings. Thanks!

This comment is part of the JOSS review: https://github.com/openjournals/joss-reviews/issues/6604
