### 2024-04-15 #6604 @kevinyamauchi
Hello everyone! I have listed my comments for the "Software paper" section of the review below. Please let me know if you have any questions or comments!

## Overview
In the manuscript, Nirmal et al. describe SCIMAP, a toolkit for single-cell analysis of multiplexed imaging data. Overall, the manuscript is well written and the described functionality is useful for researchers who want to study the spatial relationships between cells multiplexed imaging data. I think the manuscript would be strengthened by a more complete discussion of the state of the art and how SCIMAP is positioned within that landscape.

## Comments on review sections
### Summary
I think the summary is clear and well written.

### Statement of need
I think the analysis use case is nicely stated in the paper. To help motivate the need for this software and inform users if this tool is the right one for their use case, I think it would be useful to discuss where this software sits in the landscape of spatial omics tools.

### State of the field
As stated above in the "Statement of need", I think there is a need for some more in-depth discussion about how SCIMAP relates to other libraries for analyzing spatial omics data. In particular, when would a use choose SCIMAP over other libraries (and vice versa).

### Quality of writing
I think the manuscript is well-written and does not need additional editing for writing quality.

### References
Generally the citations are sufficient and the citations are rendered correctly. I noticed that a few software libraries are mentioned and not cited, but I am not sure what the JOSS policy is (e.g., if they have a zenodo DOI, but no manuscript/paper). @csoneson, would you mind clarifying? Thanks!
