### 2024-04-15 labsyspharm/scimap#100 @kevinyamauchi
How do I run the tests?
Hello! I am going through the repo for the JOSS review (https://github.com/openjournals/joss-reviews/issues/6604) and the reviewer checklist is asking about the testing infrastructure. I have a few questions below. Please let me know if anything needs clarification. Thanks!

For context, this is the prompt about testing in the reviewer checklist: "Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?"

- Are there instructions on how to run the tests locally in the documentation?
- Are the tests run on CI? I saw [this actions workflow](https://github.com/labsyspharm/scimap/blob/master/.github/workflows/build-unix-mac-win.yml) that mentions tests, but I couldn't see where in the workflow the tests are run. I am not familiar with poetry, so apologies if I missed it.
