### 2024-04-15 #6604 @kevinyamauchi
Hello again! I have listed my comments for the "Documentation" section of the review below. Please let me know if you have any questions or comments!

## Overview
Overall, the documentation for this project is very nice. There are many examples and the API documentation looks complete. I ran into some issues setting up the development environment and running the tests that I have described below.

## Comments on "Documentation" review sections

### Statement of need
The statement of need is succinctly stated on the landing page of the documentation. I think this is sufficient.

### Installation instructions

The installation instructions are clearly described on the [Getting Started](https://scimap.xyz/Getting%20Started) page.

### Example usage
There is a nice collection of tutorials notesbooks in the [Tutorials])(https://scimap.xyz/tutorials/md/spatial_biology_scimap/_ section of the documentation. I like how the notebooks are rendered with plots, etc. so that users know how the outputs are expected to look.

### Functionality documentation

The API documentation looks complete and is nicely formatted. There are a few places where

### Automated tests

I did find some tests in the repo, but I could not tell how they are run. I created an issue to discuss (https://github.com/labsyspharm/scimap/issues/100).

### Community guidelines

There is a clearly written [Contributing guide](https://scimap.xyz/contribute/) that describes how to set up the development environment and make contributions. I tried to set up the development environment using the contribution guide and received some errors. I created an issue to track (https://github.com/labsyspharm/scimap/issues/99).

Additionally, I encourage the authors to include a Code of Conduct so that potential contributors understand the expected standards of the community and how conflicts are resolved. However, based on the review checklist, this appears to be out of scope for this review, so I don't think it is required.
