### 2024-04-15 labsyspharm/scimap#99 @ajitjohnson
Dear @kevinyamauchi thank you for identifying this issue. I have updated the documentation to recommend users follow the instructions on `poetry` website for installation.

Briefly, `pipx` needs to be first installed using the instructions provided here: https://pipx.pypa.io/stable/installation/

```
# if macOS; assuming brew has been installed
brew install pipx
pipx ensurepath
```

I just tested it on macOS, and the installation went smoothly. I am unable to reproduce the error you encountered. Can I ask what operating system are you using? Thank you.
