### 2024-04-16 4617848
remove existing vscode dir
