### 2024-04-16 labsyspharm/scimap#98 @ajitjohnson
Thank you for bringing this to my attention. I have removed it and added it to the Git ignore list. I will push an updated version to pip once all the reviews have been addressed. Thank you.
