### 2024-04-16 labsyspharm/scimap#100 @ajitjohnson
Thank you @kevinyamauchi for your question.

I inferred that statement from their checklist as a way to test the package, so I provided an example dataset here: https://scimap.xyz/tutorials/md/demo_data_scimap/ This dataset can be used to follow along with the tutorials (https://scimap.xyz/tutorials/md/spatial_biology_scimap/) to test most of the functions in `scimap`.

I do have automated tests, as you have pointed out. Every time I push an update, the GitHub Actions automatically run the tests. They run on the latest versions of Windows, Linux, and macOS

Here is the latest run after addressing your other comments: https://github.com/labsyspharm/scimap/actions/runs/8697718202
