### 2024-04-16 labsyspharm/scimap#100 @kevinyamauchi
Hey @ajitjohnson ! Thanks for the quick response!

> I inferred that statement from their checklist as a way to test the package, so I provided an example dataset here: https://scimap.xyz/tutorials/md/demo_data_scimap/ This dataset can be used to follow along with the tutorials (https://scimap.xyz/tutorials/md/spatial_biology_scimap/) to test most of the functions in scimap.

Thanks, I will give this a go later this week.

>I do have automated tests, as you have pointed out. Every time I push an update, the GitHub Actions automatically run the tests. They run on the latest versions of Windows, Linux, and macOS
>
> Here is the latest run after addressing your other comments: https://github.com/labsyspharm/scimap/actions/runs/8697718202

I am sorry if I missed it, but I don't see where the tests were run in this Action. Is the pytest output visible somewhere?

When I look at the workflow file, I see where the package is installed, but it doesn't look like pytest is called. As I mentioned, I'm not super familiar with poetry. Does it automatically run tests after installation? If so, is it possible to show the output?

https://github.com/labsyspharm/scimap/blob/4617848b986dfb1a8019cbbd0a42826680dad239/.github/workflows/build-unix-mac-win.yml#L52
