+++ .github/workflows/build-unix-mac-win.yml
@@ -50,3 +50,6 @@ jobs:
         
       - name: Install project dependencies
         run: poetry install -vv || { rm -rf .poetry_cache/virtualenvs/*; poetry install -vv; }
+      
+      - name: Run tests with Poetry
+        run: poetry run pytest
