### 2024-04-20 labsyspharm/scimap#101 @jluethi
Package test coverage
Hey scimap team (@ajitjohnson)

Very nice package that you've built here! As part of the review for the JOSS paper, I looked into the tests for the package a bit. I generated the following coverage report:

```
> coverage run --source=scimap -m pytest
> coverage report
Name                                           Stmts   Miss  Cover
------------------------------------------------------------------
scimap/__init__.py                                12      2    83%
scimap/__main__.py                                 0      0   100%
scimap/cli/__init__.py                             0      0   100%
scimap/cli/_scimap_mcmicro.py                     89     89     0%
scimap/cli/test.py                                11     11     0%
scimap/external/__init__.py                        0      0   100%
scimap/helpers/__init__.py                         8      0   100%
scimap/helpers/_classify.py                       63     63     0%
scimap/helpers/addROI_omero.py                    60     52    13%
scimap/helpers/add_roi_scatter.py                 46     46     0%
scimap/helpers/animate.py                        185    176     5%
scimap/helpers/classify.py                        55     18    67%
scimap/helpers/downloadDemoData.py                28     23    18%
scimap/helpers/dropFeatures.py                    42     38    10%
scimap/helpers/merge_adata_obs.py                 48     39    19%
scimap/helpers/rename.py                          17      1    94%
scimap/helpers/scimap_to_csv.py                   48     23    52%
scimap/plotting/__init__.py                       19      0   100%
scimap/plotting/addROI_image.py                  148    131    11%
scimap/plotting/cluster_plots.py                  66     52    21%
scimap/plotting/densityPlot2D.py                  75     63    16%
scimap/plotting/distPlot.py                       87     77    11%
scimap/plotting/foldchange.py                     53     43    19%
scimap/plotting/gate_finder.py                    97     86    11%
scimap/plotting/groupCorrelation.py               89     72    19%
scimap/plotting/heatmap.py                       148    130    12%
scimap/plotting/image_viewer.py                   81     70    14%
scimap/plotting/markerCorrelation.py              96     80    17%
scimap/plotting/pie.py                            69     62    10%
scimap/plotting/spatialInteractionNetwork.py     103     87    16%
scimap/plotting/spatial_distance.py               73     64    12%
scimap/plotting/spatial_interaction.py            91     82    10%
scimap/plotting/spatial_pscore.py                 40     31    22%
scimap/plotting/spatial_scatterPlot.py            91     79    13%
scimap/plotting/stacked_barplot.py                61     49    20%
scimap/plotting/umap.py                          137    122    11%
scimap/plotting/voronoi.py                       175    163     7%
scimap/preprocessing/__init__.py                   4      0   100%
scimap/preprocessing/combat.py                    38     30    21%
scimap/preprocessing/log1p.py                     30     23    23%
scimap/preprocessing/mcmicro_to_scimap.py         83     31    63%
scimap/preprocessing/ngraph.py                    23     23     0%
scimap/preprocessing/rescale.py                  169     26    85%
scimap/tests/test_helpers.py                      23      0   100%
scimap/tests/test_preprocessing.py                19      0   100%
scimap/tests/test_tools.py                        66      0   100%
scimap/tools/__init__.py                          13      0   100%
scimap/tools/cluster.py                          153    104    32%
scimap/tools/foldchange.py                        81     21    74%
scimap/tools/phenotype_cells.py                  158     43    73%
scimap/tools/spatial_aggregate.py                 65     15    77%
scimap/tools/spatial_cluster.py                   48     39    19%
scimap/tools/spatial_count.py                     57     15    74%
scimap/tools/spatial_distance.py                  40     14    65%
scimap/tools/spatial_expression.py                97     42    57%
scimap/tools/spatial_interaction.py               88     21    76%
scimap/tools/spatial_lda.py                       83     16    81%
scimap/tools/spatial_pscore.py                    77     15    81%
scimap/tools/spatial_similarity_search.py        127    116     9%
scimap/tools/umap.py                              14     10    29%
------------------------------------------------------------------
TOTAL                                           4067   2728    33%

```

You have a very nice start with decent coverage of the tools. It's very nice that you have compact unit tests for the different functions.
I think the tools would benefit from some sort of integration test, e.g. running a whole workflow and validating its output a bit more thoroughly than checking a single item in the returned adata object.

The preprocessing & plotting subpackages have very low test coverages at the moment. It would be great if you added a few tests for the different preprocessing functions. Testing plotting is admittedly harder. But I saw that some of your plotting functions also return data, so you could test that they correctly processed the data you passed to them and return valid objects. Your plotting functions seem to be doing some significant data preprocessing afterwards and you can then check whether the data has been processed correctly by the plotting function. See also discussion here for testing plotting functions: https://stackoverflow.com/questions/27948126/how-can-i-write-unit-tests-against-code-that-uses-matplotlib
