### 2024-04-20 labsyspharm/scimap#101 @jluethi
Another small nitpick: The test suite currently generates 270 warnings. Would be good to review if there's anything critical in there
