### 2024-04-20 #6604 @jluethi
Hey all,

I've gone through all the checklist points in my review and am providing the feedback below. Feel free to reach out for clarification.

## Overview
Nirmal et al. give a concise and well-written overview of their Scimap tool in this JOSS paper. Scimap is a tool for analysing the outputs of processed multiplexed imaging data, creating plots and enabling interactive analysis that integrate the feature measurements, derived clustering & image data in an easily accessible manner.
The repository provides good documentation and excellent tutorials. The test coverage has some room for improvement though.
The paper gives a compelling overview on the importance of this analysis, but would benefit from a better description of scimap's role in the ecosystem. Importantly, it should be described better what data formats scimap can process and what tools can be used to prepare image data for scimap.

## Functionality

Scimap installs nicely and the tutorial runs without issues. The tutorial nicely shows how the functionality claimed in the paper can be applied to real data.
No specific performance claims were made.

## Documentation
The package is overall well documented. The tutorials are excellent and there is some initial test coverage with room for improvement.
The major area for improvement in my view is the description of the data modality that scimap can process. The paper gives the impression that it directly processes acquisition data. The tutorial makes it clearer that the data has been processed with tools like mcmicro. There is also some description of adaptation needed of existing datasets in respect to what features are required in the anndata table. I did not find a description of the overall required data structure though (e.g. what image formats are supported, how images need to be organized). If scimap is meant to primarily process mcmicro outputs, this should be clarified. If it's meant to process outputs from other tools, linking to or providing a description of the required data structure that those other tools would need to provide is essential.

### A statement of need
There is a clear statement of need in the readme that gives an overview of the functionality of scimap. A bit more context on the types of data it can process (/ which tools can be used to prepare data for scimap) would be helpful here as well.

### Installation instructions
A minor nitpick for the installation instructions in the README:
I recommend changing the instructions to:
`pip install "scimap[napari]"`
The quotation marks will make the installation work on zsh shells as well and not interfere with bash.

### Example usage
The tutorial is well written, based on real data and easy to follow. Well done!

### Functionality documentation
The main functions provide useful parameter documentation (but would benefit form short docstrings describing their functionality).

### Automated tests
The authors are providing some automated tests, but coverage is only partial. I gave more detailed feedback here: https://github.com/labsyspharm/scimap/issues/101

### Community guidelines
Community guidelines are provided

## Software paper
The paper is well written and makes a compelling case for the need of analysis software like scimap. It gives a good overview of the available features of scimap and the outputs users can expect when using it. It's somewhat missing comparisons to other tools in the area (tools like SpatialData: https://www.biorxiv.org/content/10.1101/2023.05.05.539647v1).
The major area that needs improvement though is a gap in the description of the analysis workflow:
The paper nicely describes & cites the multiplexed imaging acquisition methods that generate the multiplexed imaging data in its first paragraph. The processing of data is kept more vague and no tools for it are cited. It is also not clear from the paper what image modalities scimap is meant to support (image data format, required organization of images etc., see also Documentation section above). The paper would benefit from refering to concrete tools (like mcmicro) or image formats created by them (?) before diving into the analysis of these feature tables.
It would also be useful to clarify whether scimap is primarily made for 2D data and to what extent it can support 3D data. Much of the datastructures appear to be focused on 2D data (with having X & Y coordinates, but not Z coordinates), but 3D is mentioned in different places in the paper and the scimap package.
Besides these clarifications on the image processing, data modalities and image data dimensionality, the paper gives a good overview of scimap functionality and its importance.

### Summary
The summary is well written and give a good overview of the scope of scimap

### A statement of need
A good statement of need is provided. The target audience could be described a bit more (e.g. the output of which image analysis tools is scimap meant to process)

### State of the field
No comparison to other packages for the same analysis is provided.

### Quality of writing
The paper is well written. I don't see a need for additional editing here

One minor comment:
line 54: "selecting the underlying single data for further analysis" => "single cell data"?

### References
The references look good overall. One nitpick:

napari citation: the napari repo recommends citing it differently: https://github.com/napari/napari/blob/main/CITATION.cff
Maybe @kevinyamauchi could elaborate on how to best cite napari?
