### 2024-04-23 labsyspharm/scimap#101 @kevinyamauchi
Hello! My comment is also in the context of the JOSS review. I agree with @jluethi's comments. In addition, I think the existing tests can be improved. May of the tests have a form similar to the one below:

https://github.com/labsyspharm/scimap/blob/72d5ebfd1f85bc3fded26ebaab7de1514d74b7db/scimap/tests/test_tools.py#L128-L131

I think this could be improved by both testing the full values. Additionally, I'm not sure about the expected precision, but I find it a bit surprising to round to 2 decimal points. I would suggest having a look at the numpy testing suite for some utilities for doing array comparisons and float array comparisons: https://numpy.org/doc/stable/reference/routines.testing.html

Thanks!
