### 2024-04-23 #6604 @kevinyamauchi
> napari citation: the napari repo recommends citing it differently: https://github.com/napari/napari/blob/main/CITATION.cff Maybe @kevinyamauchi could elaborate on how to best cite napari?

Thanks for flagging this, @jluethi ! The preferred citation for napari is indeed the DOI in the CITATION.cff (this resolves to the most recent napari version): https://zenodo.org/doi/10.5281/zenodo.3555620
