### 2024-04-23 #6604 @kevinyamauchi
Hello everyone! I have finished my review. Overall, I think this project and software paper are a nice contribution to the spatial omics analysis ecosystem. I appreciate that the authors built upon established data standards which enhances interoperability. I have outlined some suggestions to improve the communication of how this library relates to the landscape of spatial omics analysis tools and to improve the test suite in the comments below. I think with these minor updates, the project will check all of the boxes.

I updated my comments after running the tutorials and some nice conversation with @ajitjohnson. They are linked below for convenience. Please let me know if you have any questions or comments!

- Comments on the "Software paper" section of the checklist: https://github.com/openjournals/joss-reviews/issues/6604#issuecomment-2055634781
- Comments on the Documentation section of the checklist: https://github.com/openjournals/joss-reviews/issues/6604#issuecomment-2056095188
