### 2024-04-24 c6456e2
check if plotting tests is compatable with github actions
