+++ .github/workflows/build-unix-mac-win.yml
@@ -24,18 +24,12 @@ jobs:
       fail-fast: false
       matrix:
         python-version: ['3.9', '3.10']
-        os: [ubuntu-latest, macos-latest, windows-latest]
+        os: [ubuntu-latest, macos-14, windows-latest]
     runs-on: ${{ matrix.os }}
     
     steps:
       - name: Checkout current branch
         uses: actions/checkout@v4
-
-      - name: Install dependencies on macOS
-        if: matrix.os == 'macos-latest'
-        run: |
-          brew install gettext
-          brew link --force gettext
         
       - name: Install required version of Python
         uses: actions/setup-python@v5
