+++ .github/workflows/build-unix-mac-win.yml
@@ -24,7 +24,7 @@ jobs:
       fail-fast: false
       matrix:
         python-version: ['3.9', '3.10']
-        os: [ubuntu-latest, macos-14, windows-latest]
+        os: [ubuntu-latest, macos-13, windows-latest]
     runs-on: ${{ matrix.os }}
     
     steps:
