### 2024-04-25 3e5dd74
add plotting tests
