### 2024-04-25 c0e3e07
modified plotting
