### 2024-04-25 d66259a
new test_log1p to check the entire array
