### 2024-04-25 3bd0e30
changing assert statements to check arrays
