### 2024-04-25 5e0cf11
remove plotting test
