+++ scimap/tests/test_tl.py
@@ -76,24 +76,24 @@ def test_spatial_interaction (adata):
     assert adata.uns['spatial_interaction'] is not None
     
 
+#spatial_count &
+#spatial_cluster
+def test_spatial_count (adata):
+    from scimap.tools.spatial_count import spatial_count
+    adata = spatial_count (adata, phenotype='phenotype',method='knn',radius=5)
+    # load expected data
+    loaded_data = np.load( os.getcwd() + '/scimap/tests/expected_test_values/test_spatial_count.npz')['data']
+    assert np.allclose(loaded_data, adata.uns['spatial_count'].to_numpy()), "The arrays do not match."
+    
+    # test spatial cluster
+    from scimap.tools.spatial_cluster import spatial_cluster
+    adata = spatial_cluster(adata, df_name='spatial_count')
+    # load expected data
+    loaded_data = load_pickle(os.getcwd() + '/scimap/tests/expected_test_values/test_spatial_cluster.pkl')
+    assert loaded_data == list(adata.obs['spatial_kmeans']), "The lists do not match."
+    
+
 # =============================================================================
-# #spatial_count &
-# #spatial_cluster
-# def test_spatial_count (adata):
-#     from scimap.tools.spatial_count import spatial_count
-#     adata = spatial_count (adata, phenotype='phenotype',method='knn',radius=5)
-#     # load expected data
-#     loaded_data = np.load( os.getcwd() + '/scimap/tests/expected_test_values/test_spatial_count.npz')['data']
-#     assert np.allclose(loaded_data, adata.uns['spatial_count'].to_numpy()), "The arrays do not match."
-#     
-#     # test spatial cluster
-#     from scimap.tools.spatial_cluster import spatial_cluster
-#     adata = spatial_cluster(adata, df_name='spatial_count')
-#     # load expected data
-#     loaded_data = load_pickle(os.getcwd() + '/scimap/tests/expected_test_values/test_spatial_cluster.pkl')
-#     assert loaded_data == list(adata.obs['spatial_kmeans']), "The lists do not match."
-#     
-# 
 # #spatial_lda
 # def test_spatial_lda (adata):
 #     from scimap.tools.spatial_lda import spatial_lda
@@ -139,9 +139,9 @@ def test_spatial_interaction (adata):
 # 
 # 
 # 
-# 
-# 
 # =============================================================================
+
+
 
 
 
