### 2024-04-26 6c2ead1
seperate spatial count and cluster tests
