### 2024-04-26 cb5937a
adding plot tests
