### 2024-04-26 2c44a64
remove cluster plot test
