### 2024-04-27 c10ea6f
more plotting tests
