### 2024-04-28 labsyspharm/scimap#101 @ajitjohnson
Dear @jluethi and @kevinyamauchi,

Thank you very much for your valuable suggestions. I have now incorporated tests for most functions. @jluethi, as most functions (tools) in this package are independent of each other, I was not entirely sure how an integrated workflow test is better than individual tests. Your suggestion to test the entire array was extremely helpful. I have implemented that for most tools.
