### 2024-04-28 labsyspharm/scimap#101 @ajitjohnson
```
36 passed, 16 warnings in 39.23s

Name                                           Stmts   Miss  Cover
------------------------------------------------------------------
scimap/__init__.py                                12      2    83%
scimap/external/__init__.py                        0      0   100%
scimap/helpers/__init__.py                         8      0   100%
scimap/helpers/addROI_omero.py                    60     52    13%
scimap/helpers/animate.py                        185    176     5%
scimap/helpers/classify.py                        55     21    62%
scimap/helpers/downloadDemoData.py                29     24    17%
scimap/helpers/dropFeatures.py                    42     25    40%
scimap/helpers/merge_adata_obs.py                 47     21    55%
scimap/helpers/rename.py                          17      1    94%
scimap/helpers/scimap_to_csv.py                   48     24    50%
scimap/plotting/__init__.py                       19      0   100%
scimap/plotting/addROI_image.py                  148    131    11%
scimap/plotting/cluster_plots.py                  66     52    21%
scimap/plotting/densityPlot2D.py                  82     29    65%
scimap/plotting/distPlot.py                       94     43    54%
scimap/plotting/foldchange.py                     62     24    61%
scimap/plotting/gate_finder.py                    97     86    11%
scimap/plotting/groupCorrelation.py               88     35    60%
scimap/plotting/heatmap.py                       148     64    57%
scimap/plotting/image_viewer.py                   81     70    14%
scimap/plotting/markerCorrelation.py              97     42    57%
scimap/plotting/pie.py                            78     33    58%
scimap/plotting/spatialInteractionNetwork.py     103     27    74%
scimap/plotting/spatial_distance.py               83     50    40%
scimap/plotting/spatial_interaction.py           101     58    43%
scimap/plotting/spatial_pscore.py                 47     11    77%
scimap/plotting/spatial_scatterPlot.py            97     27    72%
scimap/plotting/stacked_barplot.py                71     27    62%
scimap/plotting/umap.py                          144     61    58%
scimap/plotting/voronoi.py                       185     78    58%
scimap/preprocessing/__init__.py                   4      0   100%
scimap/preprocessing/combat.py                    38     14    63%
scimap/preprocessing/log1p.py                     30     16    47%
scimap/preprocessing/mcmicro_to_scimap.py         83     33    60%
scimap/preprocessing/rescale.py                  169     93    45%
scimap/tests/test_hl.py                           43      5    88%
scimap/tests/test_pl.py                          118      0   100%
scimap/tests/test_pp.py                           40      0   100%
scimap/tests/test_tl.py                           74      0   100%
scimap/tools/__init__.py                          13      0   100%
scimap/tools/cluster.py                          153    102    33%
scimap/tools/foldchange.py                        81     21    74%
scimap/tools/phenotype_cells.py                  158     39    75%
scimap/tools/spatial_aggregate.py                 65     15    77%
scimap/tools/spatial_cluster.py                   48     24    50%
scimap/tools/spatial_count.py                     57     15    74%
scimap/tools/spatial_distance.py                  40     14    65%
scimap/tools/spatial_expression.py                97     41    58%
scimap/tools/spatial_interaction.py               88     21    76%
scimap/tools/spatial_lda.py                       83     16    81%
scimap/tools/spatial_pscore.py                    77     15    81%
scimap/tools/spatial_similarity_search.py        127     42    67%
scimap/tools/umap.py                              14      3    79%
------------------------------------------------------------------
TOTAL                                           4094   1823    55%

```
