+++ paper/paper.bib
@@ -1,213 +1,80 @@
 
-@article{noauthor_catching_2022,
-	title = {Catching up with multiplexed tissue imaging},
-	volume = {19},
-	rights = {2022 Springer Nature America, Inc.},
-	issn = {1548-7105},
-	url = {https://www.nature.com/articles/s41592-022-01428-z},
-	doi = {10.1038/s41592-022-01428-z},
-	abstract = {Highly multiplexed tissue imaging continues to show its power for biomedical discovery. In this issue, we publish tools and guidance for implementing this class of methods and reporting subsequent results.},
-	pages = {259--259},
-	number = {3},
-	journaltitle = {Nature Methods},
-	shortjournal = {Nat Methods},
-	urldate = {2024-04-01},
-	date = {2022-03},
-	langid = {english},
-	note = {Publisher: Nature Publishing Group},
-	keywords = {Fluorescence imaging, Histology, Immunocytochemistry, Scientific community},
-	file = {Full Text PDF:C\:\\Users\\aj\\Zotero\\storage\\X8ANJRMA\\2022 - Catching up with multiplexed tissue imaging.pdf:application/pdf},
+@article{maliga_immune_2024,
+	title = {Immune Profiling of Dermatologic Adverse Events from Checkpoint Blockade Using Tissue Cyclic Immunofluorescence: A Pilot Study},
+	issn = {1523-1747},
+	doi = {10.1016/j.jid.2024.01.024},
+	shorttitle = {Immune Profiling of Dermatologic Adverse Events from Checkpoint Blockade Using Tissue Cyclic Immunofluorescence},
+	pages = {S0022--202X(24)00107--6},
+	journaltitle = {The Journal of Investigative Dermatology},
+	shortjournal = {J Invest Dermatol},
+	author = {Maliga, Zoltan and Kim, Daniel Y. and Bui, Ai-Tram N. and Lin, Jia-Ren and Dewan, Anna K. and Jadeja, Saagar and Murphy, George F. and Nirmal, Ajit J. and Lian, Christine G. and Sorger, Peter K. and {LeBoeuf}, Nicole R.},
+	date = {2024-02-14},
+	pmid = {38360200},
+	keywords = {Immunotherapy, Cutaneous adverse events, Immune checkpoint blockade, Immune-related adverse events, Tissue cyclic immunofluorescence},
 }
 
-@article{liu_analysis_2022,
-	title = {Analysis and Visualization of Spatial Transcriptomic Data},
+@article{gaglia_lymphocyte_2023,
+	title = {Lymphocyte networks are dynamic cellular communities in the immunoregulatory landscape of lung adenocarcinoma},
+	volume = {41},
+	issn = {1878-3686},
+	doi = {10.1016/j.ccell.2023.03.015},
[…]
