+++ paper/paper.md
@@ -31,7 +31,7 @@ Multiplexed imaging data are revolutionizing our understanding of the compositio
 
 A variety of methods have been introduced for high multiplexed imaging of tissues, including MxIF, CyCIF, CODEX, 4i, mIHC, MIBI, IBEX, and IMC [@angelo_multiplexed_2014; @gerdes_highly_2013; @giesen_highly_2014; @goltsev_deep_2018; @gut_multiplexed_2018; @tsujikawa_quantitative_2017; @lin_highly_2018]; although these methods differ in their implementations, all enable the collection of single-cell data on 20-100 proteins within preserved 2D and 3D tissue microenvironments. Analysis of high-plex images typically involves joining adjacent image tiles together and aligning channels across imaging cycles (stitching and registration) to generate a composite high-plex image and then identifying the positions and boundaries of individual cells via segmentation. The intensities of individual protein antigens, stains, and other detectable molecules are then quantified on a per-cell basis. This generates a “spatial feature table” (analogous to a count table in sequencing) that can be used to identify individual cell types and states; tools from spatial statistics are then used to identify how these cells are patterned in space from scales ranging from a few cell diameters (~10 µm) to several millimeters.
 
-Spatial feature tables provide the quantitative data for analysis of high-plex data but human inspection of the original image data remains essential. At the current state of the art, many of the critical morphological details in high-resolution images cannot be fully and accurately quantified. Segmentation is also subject to errors identifiable by humans, but not fully resolvable computationally [@baker_quality_2024]. As a consequence, computation of spatial features and relationships must be performed in combination with visualization of the underlying image data. Humans excel at identifying tissue features that correspond to classical histo-morphologies; they are also effective at discriminating foreground signals from variable background [@nirmal_cell_2023] using a process of “visual gating” (perception of high and low-intensity levels while visualizing an image). More generally, effective integration of visualization and computation enables nuanced interpretation of cellular organization in relation to established tissue architectures. `SCIMAP` uses the Python-based Napari [@chiu_napari_2022] image viewer to leverage these capabilities by providing a seamless interface to inspect and annotate high-plex imaging data alongside computational analysis. For example, we have implemented an image-based gating approach that allows users to visually determine the threshold that discriminates background from a true signal at both a whole-specimen and single-cell level. Users can also select specific regions of interest (ROIs) for selective or deeper analysis. This involves drawing ROIs over images (freehand or geometric) and then selecting the underlying single data for further analysis. This capability is essential for incorporating histopathological information on common tissue structures (e.g., epidermis, dermis, follicles), immune structures (e.g., secondary and tertiary lymphoid structures), tumor domains (e.g., tumor center, boundary, tumor buds), and tumor grade or stage (e.g., early lesions, invasive regions, established nodules). It also allows for excluding regions affected by significant tissue loss, folding, or artifactual staining [@baker_quality_2024]. `SCIMAP` then performs statistical and spatial analyses on individual ROIs or sets of ROIs. Spatial analysis, including the measurement of distances between cells, analysis of interaction patterns, categorization into neighborhoods, and scoring of these patterns, is crucial for elucidating the cellular communications that underpin the functional aspects of the biology being studied. `SCIMAP` offers various functions to facilitate these analyses. 
[…]
