### 2024-04-28 #6604 @ajitjohnson
\

> The main functions provide useful parameter documentation (but would benefit form short docstrings describing their functionality).

Hi @jluethi, I'm not fully understanding this part of your comment. Could you please elaborate? Thank you.
