### 2024-04-28 labsyspharm/scimap#101 @jluethi
@ajitjohnson This is looking very promising! With the addition of your tests of whole arrays, consider my wish for integration testing covered as well. As mentioned, I mostly wanted to see validation of more than existence of an output, which you appear to have nicely implemented here now.

If you want to simplify your life in finding where the coverage is still missing, I recommend using a tool like Codecov, see https://about.codecov.io/

It both shows you which lines miss coverage & has a nice github action to report test coverage of any PR submitted :)
