### 2024-04-28 #6604 @jluethi
@ajitjohnson Happy to elaborate. Let's take the example of the spatial_interaction function:

```
# Function
def spatial_interaction (adata,
                         x_coordinate='X_centroid',
                         y_coordinate='Y_centroid',
                         z_coordinate=None,
                         phenotype='phenotype',
                         method='radius',
                         radius=30,
                         knn=10,
                         permutation=1000,
                         imageid='imageid',
                         subset=None,
                         pval_method='zscore',
                         verbose=True,
                         label='spatial_interaction'):
    """Calculates spatial interaction in 2D.

This function does X. Be aware of Y. It uses algorithm Z.

Parameters:
        ...
    """
```

It's not a requirement that every function has this additional docstring of course. It's helpful for the top-level functions though, as this shows up in an IDE to describe what the function does. My approach typically is to add them to the functions I want people to use, not necessarily all the helper functions. And I try to have at least the initial summary line, only adding the extra paragraph if there are special things to be aware of.

See more documentation e.g. here: https://sphinxcontrib-napoleon.readthedocs.io/en/latest/example_google.html
