### 2024-05-06 labsyspharm/scimap#101 @kevinyamauchi
Hey @ajitjohnson. Thanks for the improvements to the testing suite! For the purpose of the JOSS review, the updates address my comments.
