### 2024-05-09 #6604 @csoneson
And sorry @kevinyamauchi, I totally missed your question above regarding citations - it's totally possible to cite software libraries (e.g. adding their Zenodo DOI, or just a website reference if there's no DOI).
