+++ scimap/preprocessing/mcmicro_to_scimap.py
@@ -18,85 +18,88 @@ import argparse
 import sys
 import pathlib
 
-def mcmicro_to_scimap (feature_table_path,
-                       remove_dna=True,
-                       remove_string_from_name=None,
-                       log=True,
-                       drop_markers=None,
-                       random_sample=None, 
-                       unique_CellId=True,
-                       CellId='CellID',
-                       split='X_centroid',
-                       custom_imageid=None,
-                       min_cells=None,
-                       verbose=True,
-                       output_dir=None):
+
+def mcmicro_to_scimap(
+    feature_table_path,
+    remove_dna=True,
+    remove_string_from_name=None,
+    log=True,
+    drop_markers=None,
+    random_sample=None,
+    unique_CellId=True,
+    CellId='CellID',
+    split='X_centroid',
+    custom_imageid=None,
+    min_cells=None,
+    verbose=True,
+    output_dir=None,
+):
+    """The function allows users to directly import the output from MCMICRO into `scimap`.
+    Parameters:
+        feature_table_path (list of str):
+            A list containing file paths to single-cell spatial feature tables. Each path corresponds to a unique image's feature table.
+
+        remove_dna (bool):
+            If set to `True`, channels identified by containing the substring 'dna' are excluded from the final dataset. This parameter is useful for omitting DNA-related features.
+
+        remove_string_from_name (str):
+            Specifies a substring to be removed from all channel names, aiding in the normalization of marker names across datasets.
+
+        log (bool):
+            If `True`, applies a log transformation (specifically log1p, which adds 1 before taking the logarithm) to the data. This is often used to normalize data distributions.
+
+        drop_markers (list of str):
+            A list of marker names to exclude from the analysis. For example, to remove specific markers, you would list them here: ["CD3D", "CD20"].
+
+        random_sample (int):
+            Specifies the number of cells to include in a subsampled dataset. This parameter is used for downsampling to a specified cell count.
+
+        CellId (str):
+            The name of the column in the input data that contains cell identifiers. This is used to track individual cells across analyses.
+
+        unique_CellId (bool):
+            Determines whether to automatically generate a unique identifier for each cell by combining the `CellId` with an `imageid`. Set to `False` to use the original `CellId` values directly. This is crucial for maintaining unique cell identifiers, especially when integrating multiple datasets.
+
+        split (str):
+            The name of the column that demarcates the boundary between quantitative marker data and additional metadata in the input CSV. Specifying this column allows for the separation of data into counts and metadata components.
+
+        custom_imageid (str):
+            Allows for the specification of a custom Image ID for each dataset. By default, the Image ID is derived from the name of the input CSV file.
+
+        min_cells (int):
+            Sets a threshold for the minimum number of cells required for an image to be included in the analysis. Images with fewer cells than this number are excluded. This is useful for filtering out datasets with sparse cellular data.
+
+        verbose (bool):
+            If set to `True`, the function will print detailed messages about its progress and the steps being executed.
+
+        output_dir (str):
+            The file path to the directory where output files will be saved. This parameter specifies the destination for any generated files.
+
+
+    Returns:
+        AnnData Object (anndata):
+            An annotated data object containing the processed single-cell spatial features, ready for downstream analysis.
+
+
+    Example:
+            ```
+
+            feature_table_path = ['/Users/aj/scimapExampleData/quantification/exemplar-001--unmicst_cell.csv',
+                                  '/Users/aj/scimapExampleData/quantification/exemplar-002--unmicst_cell.csv']
+
[…]
