+++ paper/paper.bib
@@ -1344,3 +1344,86 @@ Wouter-Michiel Vierdag docs napari - @melonora},
 	doi = {10.5281/zenodo.8115575},
 	file = {Snapshot:/Users/aj/Zotero/storage/VX4YHQCL/8115575.html:text/html},
 }
+
+@article{palla_squidpy_2022,
+	title = {Squidpy: a scalable framework for spatial omics analysis},
+	volume = {19},
+	rights = {2022 The Author(s)},
+	issn = {1548-7105},
+	url = {https://www.nature.com/articles/s41592-021-01358-2},
+	doi = {10.1038/s41592-021-01358-2},
+	shorttitle = {Squidpy},
+	abstract = {Spatial omics data are advancing the study of tissue organization and cellular communication at an unprecedented scale. Flexible tools are required to store, integrate and visualize the large diversity of spatial omics data. Here, we present Squidpy, a Python framework that brings together tools from omics and image analysis to enable scalable description of spatial molecular data, such as transcriptome or multivariate proteins. Squidpy provides efficient infrastructure and numerous analysis methods that allow to efficiently store, manipulate and interactively visualize spatial omics data. Squidpy is extensible and can be interfaced with a variety of already existing libraries for the scalable analysis of spatial omics data.},
+	pages = {171--178},
+	number = {2},
+	journaltitle = {Nature Methods},
+	shortjournal = {Nat Methods},
+	author = {Palla, Giovanni and Spitzer, Hannah and Klein, Michal and Fischer, David and Schaar, Anna Christina and Kuemmerle, Louis Benedikt and Rybakov, Sergei and Ibarra, Ignacio L. and Holmberg, Olle and Virshup, Isaac and Lotfollahi, Mohammad and Richter, Sabrina and Theis, Fabian J.},
+	urldate = {2024-05-15},
+	date = {2022-02},
+	langid = {english},
+	note = {Publisher: Nature Publishing Group},
+	keywords = {Data integration, Imaging, Software, Transcriptomics},
+	file = {Full Text PDF:/Users/aj/Zotero/storage/X2GQX7F5/Palla et al. - 2022 - Squidpy a scalable framework for spatial omics an.pdf:application/pdf},
+}
+
+@article{dries_giotto_2021,
+	title = {Giotto: a toolbox for integrative analysis and visualization of spatial expression data},
+	volume = {22},
+	issn = {1474-760X},
+	url = {https://doi.org/10.1186/s13059-021-02286-2},
+	doi = {10.1186/s13059-021-02286-2},
+	shorttitle = {Giotto},
+	abstract = {Spatial transcriptomic and proteomic technologies have provided new opportunities to investigate cells in their native microenvironment. Here we present Giotto, a comprehensive and open-source toolbox for spatial data analysis and visualization. The analysis module provides end-to-end analysis by implementing a wide range of algorithms for characterizing tissue composition, spatial expression patterns, and cellular interactions. Furthermore, single-cell {RNAseq} data can be integrated for spatial cell-type enrichment analysis. The visualization module allows users to interactively visualize analysis outputs and imaging features. To demonstrate its general applicability, we apply Giotto to a wide range of datasets encompassing diverse technologies and platforms.},
+	pages = {78},
+	number = {1},
+	journaltitle = {Genome Biology},
+	shortjournal = {Genome Biology},
+	author = {Dries, Ruben and Zhu, Qian and Dong, Rui and Eng, Chee-Huat Linus and Li, Huipeng and Liu, Kan and Fu, Yuntian and Zhao, Tianxiao and Sarkar, Arpan and Bao, Feng and George, Rani E. and Pierson, Nico and Cai, Long and Yuan, Guo-Cheng},
+	urldate = {2024-05-15},
+	date = {2021-03-08},
[…]
