### 2024-05-15 #6604 @ajitjohnson
Dear @jluethi @kevinyamauchi and @csoneson I believe I have now addressed all of your comments. Thank you so much for the constructive suggestions.

I have organized the questions into themes that were raised by both. Please let me know if I have missed anything.

**Q1: Where does this software sit in the landscape of spatial omics tools**

```
While packages such as squidpy, Giotto, and Seurat have the potential to manage multiplexed imaging data, their functionalities are primarily optimized for spatial transcriptomics data. In contrast, SCIMAP is specifically designed to address the unique requirements of multiplexed imaging data analysis, offering features such as image-based visual gating and the integration of prior knowledge for cellular phenotyping, among others.
```

**Q2: Described better what data formats scimap can process and what tools can be used to prepare image data for scimap.**

```
SCIMAP operates on segmented single-cell data derived from imaging data using tools such as cellpose or MCMICRO. The essential inputs for SCIMAP are: (a) a single-cell expression matrix and (b) the X and Y coordinates for each cell. Additionally, multi-stack OME-TIFF or TIFF images can be optionally provided to enable visualization of the data analysis on the original raw images.
```

**Q3: It would also be useful to clarify whether scimap is primarily made for 2D data and to what extent it can support 3D data.**

```
All tools within the scimap package for spatial analysis are compatible with both 2D and 3D data.
```

**Q4: The main functions provide useful parameter documentation (but would benefit form short docstrings describing their functionality)**

```
@jluethi Thank you for clarifying your comment. Initially, I did not include the short descriptions in the docstrings because I had written them at the top of each script. To address your feedback, I attempted to add these descriptions into the docstrings. However, this approach disrupted the documentation page due to the strict formatting required by mkdocstrings for auto-generating the documentation. Adding the short descriptions within the docstrings unfortunately breaks the page rendering. If you believe the short descriptions at the top of each script are insufficient, I could include them as comments right below the docstrings. -- please let me know.

```
