### 2024-05-16 #6604 @jluethi
Dear @ajitjohnson
Thank you for those clarifications. The details on where scimap sits in the ecosystem and what kind of data it can process are indeed very helpful! If this context has also been added in the manuscript, then 👍🏻 from my side on the overall review.

On point 4: There is probably some way to configure mkdocs to handle this, as it's a standard convention to include such information (see https://sphinxcontrib-napoleon.readthedocs.io/en/latest/example_google.html). It's not essential for this publication though and more a minor side-comment :)
