### 2024-05-16 #6604 @jluethi
@csoneson Is an updated version of the manuscript available somewhere to cross-check? I went through the checklist and all my points on the repo & documentation have been addressed. I've not been able to check for the manuscript yet.
