### 2024-05-16 #6604 @jluethi
Thanks @csoneson !

I read through the new text in the paper and I really like how you're putting it into context & describe the requirements now @ajitjohnson

I confirm that all items on the checklist are now fulfilled and I can recommend the scimap paper for publication.
