### 2024-05-17 #6604 @ajitjohnson
@jluethi and @kevinyamauchi,

Once again, thank you so much for your constructive feedback and reviews. I have enjoyed this process, and I believe more journals should follow this approach. @jluethi, I will definitely look into adding the summary to the docstrings—I’ve been running against one deadline after another for grants. Life of a new PI!
