+++ scimap/helpers/classify.py
@@ -1,175 +1,0 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-# Created on Mon Oct 26 12:04:17 2020
-# @author: Ajit Johnson Nirmal
-"""
-!!! abstract "Short Description"
-    `sm.hl.classify`: This utility function enables users to annotate cells by assessing 
-    the presence or absence of specific markers. It offers flexibility to apply classifications 
-    across the entire dataset or within previously defined subsets, such as phenotyped or 
-    clustered cell groups, facilitating targeted analyses based on marker expression.
-
-## Function
-"""
-
-
-#Library
-import pandas as pd
-import numpy as np
-
-
-# Functions
-def classify (adata, 
-              pos=None, 
-              neg=None, 
-              classify_label='passed_classify', 
-              failed_label='failed_classify',
-              phenotype=None,
-              subclassify_phenotype=None,
-              threshold = 0.5,
-              collapse_failed=True,
-              label="classify",
-              showPhenotypeLabel=False,
-              verbose=True):
-    
-    """
-Parameters:
-        adata (anndata.AnnData):  
-            The annotated data matrix for classification.
-            
-        pos (list, optional):  
-            Markers that should be expressed in the cells of interest.
-            
-        neg (list, optional):  
-            Markers that should not be expressed in the cells of interest.
-            
-        classify_label (str, optional):  
-            Label for cells that meet the classification criteria.
-            
-        failed_label (str, optional):  
-            Label for cells that do not meet the classification criteria.
-            
-        phenotype (str, required if subclassify_phenotype or collapse_failed is used):  
-            Column in `adata.obs` containing the phenotype information.
-            
-        subclassify_phenotype (list, optional):  
-            Phenotypes within which classification should be performed.
-            
-        threshold (float, optional):  
-            Threshold for determining positive or negative expression.
-            
-        collapse_failed (bool, optional):  
-            If True, unclassified cells are grouped under a single failed label.
-            
-        label (str, optional):  
-            Key under which classification results are stored in `adata.obs`.
-            
-        showPhenotypeLabel (bool, optional):  
-            If True, appends classification status to existing phenotype labels in the results.
-            
-        verbose (bool, optional):  
-            If True, prints progress and informational messages during the classification process.
-
-Returns:
-        adata (anndata.AnnData):  
-            The input AnnData object, updated with classification results in `adata.obs[label]`.
-
-Example:
-    ```python
-    
-    # Basic classification with positive and negative markers
-    adata = sm.hl.classify(adata, pos=['CD3D', 'CD8A'], neg=['PDGFRB'], label='T_cell_classification')
-    
-    # Classify specific phenotypes, preserving original phenotype labels for unclassified cells
-    adata = sm.hl.classify(adata, pos=['CD19'], neg=['CD3D'], subclassify_phenotype=['B cells'],
-                     phenotype='cell_type', collapse_failed=False, label='B_cell_subclassification')
-
-    # Use showPhenotypeLabel to append classification status to existing phenotype labels
-    adata = sm.hl.classify(adata, pos=['CD34'], neg=['CD45'], phenotype='cell_type',
-                     showPhenotypeLabel=True, label='stem_cell_classification', verbose=True)
-    
-    ```
-    """
-    
-    # clean the input
-    if isinstance(pos, str):
-        pos = [pos]
-    if isinstance(neg, str):
-        neg = [neg]
-    if isinstance(subclassify_phenotype, str):
-        subclassify_phenotype = [subclassify_phenotype]
[…]
