### 2024-05-22 #6604 @ajitjohnson
Dear @csoneson,

Thank you for facilitating this review process.

I just wanted to send a gentle reminder that it has been over six weeks since the manuscript was sent out for review. I understand that the reviewers are busy; however, if we do not receive any update from Reviewer 3, I am concerned we will be waiting for an extended period.

I appreciate your assistance and look forward to facilitating the timely progression of the manuscript. Thank you very much.
