### 2024-05-25 #6604 @csoneson
@ajitjohnson I have gone through the submission and overall it looks very good! I have made a [PR](https://github.com/labsyspharm/scimap/pull/104) with some small edits - please merge if you agree, and then generate a new proof with `@editorialbot generate pdf`.

I will make another post outlining the next steps.
