+++ pyproject.toml
@@ -1,7 +1,7 @@
 [tool.poetry]
 
 name = "SCIMAP"
-version = "2.0.5"
+version = "2.1.0"
 description = "Spatial Single-Cell Analysis Toolkit"
 
 license = "MIT"
