### 2024-05-27 #6604 @csoneson
Thanks @ajitjohnson. A couple of points:

- it looks like the PR with the minor fixes to the paper was not merged (they are not in the proof above)?
- please make sure that the title of the Zenodo archive is the same as that of the paper (edit the metadata of the Zenodo archive)
