### 2024-05-28 #6604 @csoneson
Thank you @ajitjohnson. I think we would really prefer if the title of the Zenodo archive just contains the paper title (i.e., without the ` - JOSS Paper Release - Zenodo Archive` part; the information about the release version and the fact that it's a Zenodo archive is available elsewhere on the archive page). Apart from that, all looks good.
