+++ docs/index.md
@@ -28,7 +28,7 @@ SCIMAP operates on segmented single-cell data derived from imaging data using to
 Check out other tools from the [Nirmal Lab](https://nirmallab.com/tools/). 
 
 ### Citing scimap
-Nirmal et al., (2024). SCIMAP: A Python Toolkit for Integrated Spatial Analysis of Multiplexed Imaging Data. Journal of Open Source Software, 9(97), 6604, [https://doi.org/10.21105/joss.06604](https://joss.theoj.org/papers/10.21105/joss.06604#)
+Nirmal et al., (2024). SCIMAP: A Python Toolkit for Integrated Spatial Analysis of Multiplexed Imaging Data. *Journal of Open Source Software*, 9(97), 6604, [https://doi.org/10.21105/joss.06604](https://joss.theoj.org/papers/10.21105/joss.06604#)
 
 
 ### Funding
